requirements.txt
tqdm==4.31.1
torchtext==0.3.1
tensorboardX==1.6
pyrouge==0.1.3
pytorch_pretrained_bert==0.6.2
pyspark==2.4.3
pyarrow==0.13.0
